+++ CHANGES.md
@@ -28,6 +28,7 @@ Change Log
 * Added new construction options to `CesiumWidget` and `Viewer`, for `skyBox`, `skyAtmosphere`, and `globe`.
 * Fixed a bug that prevented Cesium from working in browser configurations that explicitly disabled localStorage, such as Safari's private browsing mode.
 * Cesium is now tested using Jasmine 2.2.0.
+* Exception will be thrown on updating Primitive.modelMatrix in 2D or Columbus View or if it has more than one instance.
 
 ### 1.7.1 - 2015-03-06